debian 12 to debian buster fix

+++ srcs/requirements/wordpress/tools/setup.sh
@@ -4,31 +4,37 @@ sleep 10
 
 cd /var/www/html
 
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+if [ ! -f /usr/local/bin/wp ]; then
+    curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+    chmod +x wp-cli.phar
+    mv wp-cli.phar /usr/local/bin/wp
+fi
 
-chmod +x wp-cli.phar
-mv wp-cli.phar /usr/local/bin/wp
+if [ ! -f wp-config.php ]; then
 
-wp core download --allow-root
+    wp core download --allow-root
 
-wp config create \
---dbname=$MYSQL_DATABASE \
---dbuser=$MYSQL_USER \
---dbpass=$MYSQL_PASSWORD \
---dbhost=mariadb \
---allow-root
+    wp config create \
+        --dbname=$MYSQL_DATABASE \
+        --dbuser=$MYSQL_USER \
+        --dbpass=$MYSQL_PASSWORD \
+        --dbhost=mariadb \
+        --allow-root
 
-wp core install \
---url=$DOMAIN_NAME \
---title="Inception" \
---admin_user=$WP_ADMIN_USER \
---admin_password=$WP_ADMIN_PASSWORD \
---admin_email=$WP_ADMIN_EMAIL \
---skip-email \
---allow-root
+    wp core install \
+        --url=$DOMAIN_NAME \
+        --title="Inception" \
+        --admin_user=$WP_ADMIN_USER \
+        --admin_password=$WP_ADMIN_PASSWORD \
+        --admin_email=$WP_ADMIN_EMAIL \
+        --skip-email \
+        --allow-root
 
-wp user create $WP_USER $WP_USER_EMAIL \
---user_pass=$WP_USER_PASSWORD \
---allow-root
+    wp user create $WP_USER $WP_USER_EMAIL \
+        --user_pass=$WP_USER_PASSWORD \
+        --allow-root
+fi
 
-exec php-fpm8.2 -F
+mkdir -p /run/php
+
+exec php-fpm7.3 -F
